Hotel_Project/requirements.txt
django==3.0.5
django-allauth==0.42.0
django-autoslug==1.9.8
django-import-export==2.3.0
django-embed-video==1.3.3
pycryptodome
pillow
requests==2.23.0
razorpay